backend/requirements.txt
Django==5.0.1
django-environ==0.11.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
ldap3==2.9.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
PyJWT==2.8.0
django-filter==23.5
django-guardian==2.4.0
django-simple-history==3.4.0
django-storages==1.14.2
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-redis==5.4.0
redis==5.0.1
celery==5.3.6
flower==2.0.1
drf-yasg==1.21.7